requirements.txt
flask
flask-cors
gunicorn
sentence-transformers
faiss-cpu
pandas
numpy
requests
rapidfuzz
unidecode
